Added addRecipe method + renaming

+++ src/main/java/com/example/blps_lab1/repository/IngredientsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IngredientsRepository extends JpaRepository<Ingredients, Integer> {
+public interface IngredientsRepository extends JpaRepository<Ingredients, Long> {
 }
